backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# RAG & AI
langchain==0.1.0
langchain-community==0.0.12
chromadb==0.4.22
sentence-transformers==2.2.2
groq==0.4.2
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
